requirements.txt
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
joblib==1.3.2

streamlit==1.32.2

jupyterlab==4.0.9
notebook==7.0.6
ipykernel==6.29.0

spacy==3.7.2
pdfplumber==0.10.3
regex==2023.12.25
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0

langchain==0.2.14
langgraph==0.2.16
crewai==0.41.1
openai==1.40.6
tiktoken==0.7.0
pydantic==2.7.4
python-dotenv==1.0.1
langchain-openai==0.1.22

fastapi==0.110.0
uvicorn==0.29.0
Flask==2.3.3
Werkzeug==2.3.8
Flask-Cors==4.0.0

xgboost==1.7.6
